python-tool/requirements.txt
Pillow>=10.0.0
pillow-heif>=0.13.0
pikepdf>=8.0.0
